Split out thread_proxy.h
If we have it in its own file then we can have a nice header comment
which will hopefully make some sense.
